Initial codemod to rename import sources. Not working.
Trying to modify and return an element in replaceWith is not working.
I need either to find a api tool other than replaceWith to modify with,
or I may need to construct a fresh ImportDeclaration to replace the original
with (but I'm worried about not knowing what kind of context would be good
to attach to that generated ImportDeclaration).